modern_requirements.txt
aiogram==3.4.1
aiomysql==0.2.0
qrcode[pil]==7.4.2
cryptography==42.0.0
python-dotenv==1.0.0
Pillow==10.2.0
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.2
uvloop==0.19.0
PyMySQL==1.1.0
